+++ database/migrations/2016_02_22_055555_create_blocks_tables.php
@@ -12,25 +12,27 @@ class CreateBlocksTables extends Migration
      */
     public function up()
     {
-        Schema::create('blocks', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('symbol_id')->unsigned();
-            $table->foreign('symbol_id')->references('id')->on('symbols')->onDelete('no action');
-            $table->integer('exchange_id')->unsigned();
-            $table->string('discount');
-            $table->string('number_shares');
-            $table->string('discount_target');
-            $table->integer('need_id')->unsigned();
-            $table->foreign('need_id')->references('id')->on('needs')->onDelete('no action');
-            $table->integer('source_id')->unsigned();
-            $table->foreign('source_id')->references('id')->on('sources')->onDelete('no action');
-            $table->integer('custodian_id')->unsigned();
-            $table->foreign('custodian_id')->references('id')->on('custodians')->onDelete('no action');
-            $table->integer('rep_id')->unsigned();
-            $table->foreign('rep_id')->references('id')->on('reps')->onDelete('no action');
-            $table->enum('status', array('unpublished','published', 'archived'));
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('blocks')) {
+            Schema::create('blocks', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('symbol_id')->unsigned();
+                //$table->foreign('symbol_id')->references('id')->on('symbols')->onDelete('no action');
+                $table->integer('exchange_id')->unsigned();
+                $table->string('discount');
+                $table->string('number_shares');
+                $table->string('discount_target');
+                $table->integer('need_id')->unsigned();
+                $table->foreign('need_id')->references('id')->on('needs')->onDelete('no action');
+                $table->integer('source_id')->unsigned();
+                $table->foreign('source_id')->references('id')->on('sources')->onDelete('no action');
+                $table->integer('custodian_id')->unsigned();
+                $table->foreign('custodian_id')->references('id')->on('custodians')->onDelete('no action');
+                $table->integer('rep_id')->unsigned();
+                $table->foreign('rep_id')->references('id')->on('reps')->onDelete('no action');
+                $table->enum('status', array('unpublished','published', 'archived'));
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -40,6 +42,8 @@ class CreateBlocksTables extends Migration
      */
     public function down()
     {
-        Schema::drop('blocks');
+        if (Schema::hasTable('blocks')) {
+            Schema::drop('blocks');
+        }
     }
 }
